+++ public/admin/_app/immutable/nodes/41.D3WGb5wM.js
@@ -1,1 +1,0 @@
-import{It as e}from"../chunks/DFnvYPbU.js";import{c as t,l as n}from"../chunks/3TA2i3Oq.js";function r(e,t){throw new n(e,t.toString())}var i=e({load:()=>a}),a=()=>{throw r(308,`${t}/settings/users`)};export{i as universal};